refactor adaptive max pool
